Finish reify literal, refactoring, basic test works
