+++ grammars/silver/translation/java/core/ProductionDcl.sv
@@ -127,13 +127,8 @@ ${implode("", map(makeChildAccessCaseLazy, namedSig.inputElements))}
 ${body.translation}
 	}
 	
-	static ${className} reify(final common.BaseTypeRep resultType, final java.util.List<core.reflect.NAST> childASTs, final java.util.Map<String, core.reflect.NAST> annotationASTs) {
-		${implode(
-		  "\n\t\t",
-		  map(
-		    \ tv::TyVar ->
-		      s"common.VarTypeRep typeVar_${toString(tv.extractTyVarRep)} = new common.VarTypeRep();",
-		    namedSig.typerep.freeVariables))}
+	public static ${className} reify(final common.TypeRep resultType, final java.util.List<core.reflect.NAST> childASTs, final java.util.Map<String, core.reflect.NAST> annotationASTs) {
+		${makeTyVarDecls(namedSig.typerep.freeVariables)}
 		
 		if (!resultType.unify(${namedSig.outputElement.typerep.transTypeRep}, true)) {
 			throw new common.exceptions.SilverError("reify is constructing " + resultType.toString() + ", but found ${ntName} AST (production ${fName}).");
@@ -177,6 +172,19 @@ ${body.translation}
     if id.name == "main"
     then [err(top.location, "main should be a function!")]
     else [];
+}
+
+function makeTyVarDecls
+String ::= vars::[TyVar]
+{
+  return
+    implode(
+      "\n\t\t",
+      map(
+        \ tv::TyVar ->
+          s"common.VarTypeRep typeVar_${toString(tv.extractTyVarRep)} = new common.VarTypeRep();",
+          vars));
+    
 }
